simplify privacy practices - it only ever loads once so it doesn't need to rerender
