+++ shared/js/ui/templates/privacy-practices.es6.js
@@ -1,6 +1,7 @@
 const bel = require('bel')
+const changeCase = require('change-case')
 const hero = require('./shared/hero.es6.js')
-const details = require('./shared/privacy-practices-details.es6.js')
+const statusList = require('./shared/status-list.es6.js')
 
 module.exports = function () {
   const domain = this.model && this.model.domain
@@ -27,11 +28,46 @@ module.exports = function () {
       </div>
       <div class="privacy-practices__details padded border--bottom--inner
           js-privacy-practices-details">
-        ${details(tosdr)}
+        ${tosdr && tosdr.reasons ?  renderDetails(tosdr.reasons) : renderNoDetails()}
       </div>
       <div class="privacy-practices__attrib padded text--center">
         Privacy Practice results from <a href="https://tosdr.org/" class="bold" target="_blank">ToS;DR</a>
       </div>
     </div>
   </section>`
+}
+
+function renderDetails (reasons) {
+  let good = reasons.good || []
+  let bad = reasons.bad || []
+
+  if (!good.length && !bad.length) return renderNoDetails()
+
+  // convert arrays to work for the statusList template,
+  // which use objects
+
+  good = good.map(item => ({
+    msg: changeCase.upperCaseFirst(item),
+    modifier: 'good'
+  }))
+
+  bad = bad.map(item => ({
+    msg: changeCase.upperCaseFirst(item),
+    modifier: 'bad'
+  }))
+
+  // list good first, then bad
+  return statusList(good.concat(bad))
+}
+
+function renderNoDetails () {
+  return bel`<div class="text--center">
+    <div class="privacy-practices__details__no-detail-icon"></div>
+    <h1 class="privacy-practices__details__title">
+      No Privacy Practices Found
+    </h1>
+    <div class="privacy-practices__details__msg">
+      The Privacy practices of this website have not been reviewed
+    </div>
+  </div>`
 }
